Small performance improvement
```
name                   old time/op    new time/op    delta
LoadTransitionFiles-4     354µs ± 1%     351µs ± 1%  -0.87%  (p=0.001 n=15+15)

name                   old alloc/op   new alloc/op   delta
LoadTransitionFiles-4    54.6kB ± 0%    54.5kB ± 0%  -0.26%  (p=0.000 n=14+14)

name                   old allocs/op  new allocs/op  delta
LoadTransitionFiles-4       942 ± 0%       933 ± 0%  -0.96%  (p=0.000 n=15+15)
```