@nicksnyder 2017-06-17
This change is backwards incompatible and it looks like the perf gain isn't significant, so I am going to leave this PR open until there are other backward incompatible changes being made.