@bogem 2017-06-18
Yes, it successfully compiles.
``` go
$ cat lang.go
package main

import "github.com/nicksnyder/go-i18n/i18n/language"

func main() {
	_ = language.Language{"ru", language.EmptyPluralSpec}
}

$ go run lang.go

```